Tweak make_day.sh

+++ make_day.sh
@@ -2,8 +2,6 @@
 
 set -eu
 
-echo "$#"
-
 if [ $# -ne 2 ]; then
   echo "Usage: $0 YEAR DAY"
   exit 1
@@ -15,16 +13,22 @@ PADDED_DAY=$(printf "%02d" $DAY)
 
 mkdir -p "inputs/$YEAR"
 
-curl -f --cookie "session=$AOC_SESSION" "https://adventofcode.com/$YEAR/day/$DAY/input" > "inputs/$YEAR/$PADDED_DAY.txt"
-echo "downloaded input"
+printf "downloading input..."
+
+if ! curl -f --cookie "session=$AOC_SESSION" "https://adventofcode.com/$YEAR/day/$DAY/input" > "inputs/$YEAR/$PADDED_DAY.txt" 2>/dev/null; then
+  echo "failed"
+  exit 1
+fi
+
+echo "downloaded"
 
 SRC_RS="src/y$YEAR/day$PADDED_DAY.rs"
 
+printf "copy template..."
 if [ -e $SRC_RS ]; then
     echo "$SRC_RS already exists"
     exit 1
 fi
 
-printf "copy template..."
 cp dayNN.rs.tpl src/y$YEAR/day$PADDED_DAY.rs
 echo "done"
